requirements.txt
# Web Frameworks (updated versions)
fastapi>=0.100.0
uvicorn[standard]>=0.27.0
streamlit>=1.29.0

# Database
sqlalchemy>=2.0.0

# Data Validation (updated for Python 3.13)
pydantic>=2.10.0
pydantic-settings>=2.0.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Task Scheduling
APScheduler>=3.10.0

# Image Processing (Python 3.13 compatible)
Pillow>=11.0.0

# AI APIs
google-generativeai>=0.3.0

# Visualization
plotly>=5.18.0

# Development
pytest>=7.4.0
httpx>=0.25.0
email-validator>=2.1.0

# Date Handling
python-dateutil>=2.8.0
